web interface for floppy selection works, kinda
The only excuse for this shitty commit is to get a working demo for CC'2018.

 * a bit more robust state handling in the esp8266
 * timeouts on both sides, so that nothing lingers forever hopefully
 * delayed spi slave initialisation seems to help with reboots on bootup
 * floppy/spi.v works at 24MHz for sd card, at 12MHz for ESP12F

Checking this in because there's no time to make it better. In the
future, the protocol between SoC and ESP should be changed. Both SoC and
ESP8266 sides are insanely sloppy. This is only for the demo.

Why the fuck ESP keeps resetting is beyond my understanding. It's just
shit. It helps to initialise spi slave a bit later in a timer callback.
If it's initialised in the user_main(), the SoC starts polling it before
something else important happens inside of the ESP and it just keeps on
rebooting.

When there's time:
 * the SoC side needs a proper co-routine to avoid getting stuck in the loop
 * ESP side also should be rewritten using callbacks and HTTPD_GET_MORE
 or what's that thing called
 * spi.v should be split into two separate modules